debug:  Allow MWC 2.1 without __PRETTY_FUNCTION__
This code was brokenly checking for Metrowerks C++ 0.3, not 3.0, but 2.4
supports __PRETTY_FUNCTION__ anyway.  If it turns out that 2.2 does not,
then 0x2200 needs to be raised to either 0x2300 or 0x2400, (depending on
whether 2.3 does or doesn't).  It's also possible that 2.2 or 2.3 allows
__func__ but not __PRETTY_FUNCTION__, in which case that condition needs
to be added.

If neither of __func__ or __PRETTY_FUNCTION__ is supported, fall back to
"(unknown)".